+++ assets/components/widgets/compareRace/compareRace.jsx
@@ -154,9 +154,13 @@ UI.widgets.CompareRaceDriver = React.createClass({
 				}
 				</div>
 				<div className="compareName"><div className="name">{self.fixName(driver.name)}</div></div>
+				{UI.state.controllerOptions.options.showLiveryPreview.value === "true" ?
 				<div className="vehicle">
 					<img src={'http://game.raceroom.com/store/image_redirect?id='+driver.liveryId+'&size=small'} />
 				</div>
+				:
+					null 
+				}
 				<div className="manufacturer">
 					<img src={'/render/'+driver.manufacturerId+'/small/'}/>
 				</div>